+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `hasListPrice`  on  `product-selling-price-range`.
+
 ## [1.28.0] - 2022-06-01
 
 ### Added